Move a bunch of selectors from product-details to products.
